+++ language/src/main/java/i4gl/nodes/cast/CastToVarcharNode.java
@@ -28,14 +28,19 @@ public abstract class CastToVarcharNode extends UnaryNode {
         return argument;
     }
 
+    @Specialization
+    Varchar castText(String argument) {
+        Varchar value = (Varchar) getVarcharType().getDefaultValue();
+        value.assignString(argument);
+        return value;
+    }
+
     @Specialization(guards = "inputs.isString(argument)", limit = "2")
-    Varchar castText(Object argument, @CachedLibrary("argument") InteropLibrary inputs) {
+    Varchar castObject(Object argument, @CachedLibrary("argument") InteropLibrary inputs) {
         try {
             String asText = inputs.asString(argument);
-            Varchar value = (Varchar) getVarcharType().getDefaultValue();
-            value.assignString(asText);
-            return value;
-        } catch (UnsupportedMessageException e) {
+            return castText(asText);
+         } catch (UnsupportedMessageException e) {
             CompilerDirectives.transferToInterpreterAndInvalidate();
             throw new InvalidCastException(argument, getVarcharType());
         }